Add new `max_last_will_delay` value and change default
The current default of `0s` was confusing to the users and as long as
it is possible from the broker side to limit this then it makes sense
for the default to be what the client provides. This can now be done
by specifying `client` instead of a time value.